format and cleanups
